--- orga/app/components/import.hbs
@@ -1,58 +1,0 @@
-<article class="import-students-page">
-  <header class="import-students-page__header">
-    <h1>
-      {{t this.textsByOrganizationType.title}}
-    </h1>
-  </header>
-
-  <ImportBanner
-    @organizationImportDetail={{@organizationImportDetail}}
-    @errorPanelId="import-error-messages"
-    @isLoading={{@isLoading}}
-  />
-
-  {{#if @isLoading}}
-    <Ui::PixLoader />
-  {{/if}}
-
-  {{#unless @isLoading}}
-    <div class="import-students-page__type-list">
-      {{#if this.currentUser.isSCOManagingStudents}}
-        <ScoOrganizationParticipant::Add
-          @importHandler={{@onImportScoStudents}}
-          @supportedFormats={{this.supportedFormats}}
-          @disabled={{this.inProgress}}
-        />
-      {{else if this.currentUser.isSUPManagingStudents}}
-        <SupOrganizationParticipant::ImportCards::Add
-          @importHandler={{@onImportSupStudents}}
-          @supportedFormats={{this.supportedFormats}}
-          @disabled={{this.inProgress}}
-        />
-        <SupOrganizationParticipant::ImportCards::Replace
-          @importHandler={{@onReplaceStudents}}
-          @supportedFormats={{this.supportedFormats}}
-          @disabled={{this.inProgress}}
-        />
-      {{else if this.currentUser.hasLearnerImportFeature}}
-        <ScoOrganizationParticipant::Add
-          @importHandler={{@onImportLearners}}
-          @supportedFormats={{this.supportedFormats}}
-          @disabled={{this.inProgress}}
-        />
-      {{/if}}
-    </div>
-  {{/unless}}
-
-  {{#if this.displayImportMessagePanel}}
-    <section id="import-error-messages" class={{this.panelClasses}} aria-live="assertive">
-      <h2>{{t "pages.organization-participants-import.error-panel.title"}}</h2>
-
-      <ul class="import-students-page__error-panel-list">
-        {{#each this.errorDetailList as |errorElement|}}
-          <li class="import-students-page__error-panel-list__item">{{errorElement}}</li>
-        {{/each}}
-      </ul>
-    </section>
-  {{/if}}
-</article>
