--- orga/app/components/sup-organization-participant/import-cards/replace.hbs
@@ -1,22 +1,0 @@
-<ImportCard @cardTitle={{t "pages.organization-participants-import.actions.replace.title"}}>
-  <:cardDetails> {{t "pages.organization-participants-import.actions.replace.details"}}</:cardDetails>
-  <:cardFooter>
-    <PixButton
-      @triggerAction={{this.toggleModal}}
-      @isDisabled={{@disabled}}
-      @size="small"
-      aria-describedby="accepted-files-replace"
-    >
-      {{t "pages.organization-participants-import.actions.replace.label"}}
-    </PixButton>
-    <p class="import-card__accepted-files" id="accepted-files-replace">
-      {{this.acceptedFileType}}
-    </p>
-  </:cardFooter>
-</ImportCard>
-<SupOrganizationParticipant::Modal::ReplaceStudentsModal
-  @onClose={{this.toggleModal}}
-  @display={{this.displayModal}}
-  @onReplaceStudents={{@importHandler}}
-  @supportedFormats={{@supportedFormats}}
-/>
