--- orga/app/components/sup-organization-participant/modal/replace-students-modal.hbs
@@ -1,48 +1,0 @@
-<PixModal
-  @title={{t "pages.sup-organization-participants.replace-students-modal.title"}}
-  @showModal={{@display}}
-  @onCloseButtonClick={{@onClose}}
->
-  <:content>
-    <div class="replace-students-modal">
-      <p class="replace-students-modal__content">
-        {{t "pages.sup-organization-participants.replace-students-modal.main-content"}}
-      </p>
-      <p class="replace-students-modal__content">
-        {{t "pages.sup-organization-participants.replace-students-modal.footer-content"}}
-      </p>
-      <p class="replace-students-modal__content">
-        <strong>{{t "pages.sup-organization-participants.replace-students-modal.last-warning"}}</strong>
-      </p>
-      <PixCheckbox {{on "click" this.giveDeletionPermission}} @size="small" @checked={{this.allowDeletion}}>
-        <:label><strong>{{t
-              "pages.sup-organization-participants.replace-students-modal.confirmation-checkbox"
-            }}</strong></:label>
-      </PixCheckbox>
-    </div>
-  </:content>
-  <:footer>
-    <PixButton @triggerAction={{@onClose}} @variant="secondary">
-      {{t "common.actions.cancel"}}
-    </PixButton>
-    {{#if this.allowDeletion}}
-      <PixButtonUpload
-        @id="students-file-upload-replace"
-        @onChange={{@onReplaceStudents}}
-        @variant="error"
-        accept={{@supportedFormats}}
-      >
-        {{t "pages.sup-organization-participants.replace-students-modal.confirm"}}
-      </PixButtonUpload>
-    {{else}}
-      <PixButton
-        @id="students-file-upload-replace"
-        @triggerAction={{(@onReplaceStudents)}}
-        @isDisabled={{true}}
-        @variant="error"
-      >
-        {{t "pages.sup-organization-participants.replace-students-modal.confirm"}}
-      </PixButton>
-    {{/if}}
-  </:footer>
-</PixModal>
