+++ orga/tests/integration/components/import/index_test.js
@@ -6,7 +6,7 @@ import { t } from 'ember-intl/test-support';
 import { module, test } from 'qunit';
 import sinon from 'sinon';
 
-import setupIntlRenderingTest from '../../helpers/setup-intl-rendering';
+import setupIntlRenderingTest from '../../../helpers/setup-intl-rendering';
 
 module('Integration | Component | Import', function (hooks) {
   setupIntlRenderingTest(hooks);
