+++ orga/tests/unit/components/import/index_test.js
@@ -2,9 +2,9 @@ import { setupTest } from 'ember-qunit';
 import { module, test } from 'qunit';
 import sinon from 'sinon';
 
-import createGlimmerComponent from '../../helpers/create-glimmer-component';
+import createGlimmerComponent from '../../../helpers/create-glimmer-component';
 
-module('Unit | Component | import', (hooks) => {
+module('Unit | Component | Import', (hooks) => {
   setupTest(hooks);
 
   let component;
